+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.1] - 2026-03-09
+
+### Added
+
+- `:as` option for `defvariant` to customize the generated function name. By default, variant
+  functions are named `{variant}_{base}` (e.g. `build_admin_user_struct`). The `:as` option
+  overrides this combined name (e.g. `as: :mod` generates `build_mod_struct` instead of
+  `build_moderator_user_struct`).
+
 ## [0.2.0] - 2026-03-08
 
 ### Added